feat: moved report_html under .strategy_tester.report.html module
