+++ src/strategy_tester/backtesting/__init__.py
@@ -5,6 +5,5 @@ You cand find:\n
 
 + `.pipeline` module with default pipelines and their helper methods\n
 + `run_backtest` to perform a backtest for the strategy, see `.backtest.run_backtest`\n
-+ `report_html` to perform a backtest for the strategy, see `.report_html`\n
 + `saving` module with backtest result saving functions, see `.saving`\n
 """
