+++ src/strategy_tester/report/__init__.py
@@ -3,4 +3,5 @@ Module with helper functions to make report files.
 
 Inside this module you can find:\n
 + `report_to_pdf` to save statistics, metrics and so on as pdf, see `.pdf.report_pdf`\n
++ `report_to_html` to save statistics, metrics and so on as html, see `.html.report_html`\n
 """
