+++ src/strategy_tester/report/html/__init__.py
@@ -1,0 +1,7 @@
+"""
+Module with helper functions to make report as html.
+
+Inside this module you can find:\n
++ `report_to_html` to save statistics, metrics and so on as html, see `.report_html.report_to_html`\n
+"""
+from .report_html import report_to_html
