+++ src/strategy_tester/report/html/report_html.py
@@ -1,527 +1,18 @@
 import base64
 import json
-import math
-from itertools import combinations
 
 import pandas as pd
-import matplotlib as mpl
 
-from ..pipeline.context import Context
-
-_viridis_cmap = mpl.colormaps["viridis"]
-# x = np.linspace(0.0, 1.0, 100)
-# rgb = viridis_cmap(x)[np.newaxis, :, :3]
-# print(rgb)
-_viridis_cmap_hex_colors = []
-for i in range(_viridis_cmap.N):
-  rgba = _viridis_cmap(i)
-  _viridis_cmap_hex_colors.append(mpl.colors.rgb2hex(rgba))
-
-def _flatten_list(matrix):
-  flat_list = []
-  for row in matrix:
-    flat_list += row
-  return flat_list
-
-def _dataframe_to_grouped_barchart_options(data: pd.DataFrame, title: str = None):
-  if data.empty:
-    return None
-  
-  x_values = data.index.map(lambda x: str(x)).to_list()
-  series = []
-
-  for key in data.columns:
-    df = data[[key]]
-    series.append({
-      "type": "bar",
-      "name": key,
-      "data": _flatten_list(df.values.tolist())
-    })
-
-  options = {
-    "animation": False,
-    "tooltip": {},
-    "xAxis": {
-      "data": x_values,
-    },
-    "yAxis": {},
-    "legend": {
-      "right": 50,
-    },
-    "series": series
-  }
-
-  if title:
-    options["title"] = { "text": title }
-  
-  return options
-
-def _dataseries_to_stacked_barchart_options_splitted_by_zeroline(data: pd.Series, title: str = None):
-  if data.empty:
-    return None
-  
-  x_values = data.index.map(lambda x: str(x)).to_list()
-  series = []
-
-  positives = data.map(lambda x: x if x > 0 else 0).to_list()
-  negatives = data.map(lambda x: x if x < 0 else 0).to_list()
-
-  series.append({
-    "type" : "bar",
-    "stack": "total",
-    "name": "positive",
-    "data" : positives,
-  })
-  series.append({
-    "type" : "bar",
-    "stack": "total",
-    "name": "negative",
-    "data" : negatives,
-  })
-
-  options = {
-    "animation": False,
-    "tooltip": {},
-    "xAxis"  : {
-      "data": x_values,
-    },
-    "yAxis" : {},
-    "legend": {
-      "right": 50,
-    },
-    "series": series
-  }
-
-  if title:
-    options["title"] = { "text": title }
-
-  return options
-
-def _dataseries_to_linechart_options(data: pd.Series, title: str = None):
-  if data.empty:
-    return None
-  
-  x_values = data.index.map(lambda x: str(x)).to_list()
-  series = [{ "type": "line", "data": data.values.tolist() }]
-  options = {
-    "animation": False,
-    "tooltip": {},
-    "xAxis": {
-      "data": x_values,
-    },
-    "yAxis": {},
-    "legend": {},
-    "series": series
-  }
-
-  if title:
-    options["title"] = { "text": title }
-
-  return options
-
-def _dataframe_to_scatterchart_cartesian_options(data: pd.DataFrame, columns, title: str = None):
-  if data.empty:
-    return None
-  
-  series = [{ "type": "scatter", "data": data[columns].values.tolist() }]
-  options = {
-    "animation": False,
-    "tooltip": {},
-    "xAxis": {},
-    "yAxis": {},
-    "series": series
-  }
-
-  if title:
-    options["title"] = { "text": title }
-
-  return options
-
-def _dataframe_to_candlestickchart_options(ohlcv: pd.DataFrame, trades: pd.DataFrame):
-  if ohlcv.empty or trades.empty:
-    return None
-  
-  df = ohlcv.copy().reset_index()
-  df[df.columns[0]] = df[df.columns[0]].map(lambda x: str(x))
-  dataset = df.values.tolist()
-  trades_markers = []
-  trades_lines = []
-
-  for i, row in trades.iterrows():
-    position_size = row["Size"]
-    pnl = row["PnL"]
-    entry_time = str(row["EntryTime"]) # trades["EntryTime"][i].replace(' ', '\n');
-    exit_time = str(row["ExitTime"]) # trades["ExitTime"][i].replace(' ', '\n');
-    entry_row = ohlcv[ohlcv.index == entry_time]
-    exit_row = ohlcv[ohlcv.index == exit_time]
-    entry_marker_price = float((entry_row["Low"] if position_size > 0 else entry_row["High"]).values[0])
-    exit_marker_price = float((exit_row["High"] if position_size > 0 else exit_row["Low"]).values[0])
-    entry_marker = {
-      "name": f"Trade {i}",
-      "symbol": "triangle",
-      "symbolSize": [20, 30],
-      "symbolRotate": 0 if position_size > 0 else 180,
-      "coord": [entry_time, entry_marker_price],
-      "value": "E",
-      "itemStyle": {
-        "color": "green" if position_size > 0 else "red"
-      }
-    }
-    exit_marker = {
-      "name": f"Trade {i}",
-      "symbol": "diamond",
-      "symbolSize": [30, 30],
-      "symbolRotate": 180 if position_size > 0 else 0,
-      "coord": [exit_time, exit_marker_price],
-      "value": position_size,
-      "itemStyle": {
-        "color": "green" if pnl > 0 else "red"
-      }
-    }
-    trades_markers.append(entry_marker)
-    trades_markers.append(exit_marker)
-
-    trade_line_entry = {
-      "coord": entry_marker["coord"],
-      "symbol": "none",
-      "value": round(pnl, 2),
-      "lineStyle": {
-        "color": "green" if pnl > 0 else "red",
-        "type": "dashed",
-        "dashOffset": 5,
-      },
-    }
-    trade_line_exit = {
-      "coord": exit_marker["coord"],
-      "symbol": "none",
-      "value": round(pnl, 2),
-      "lineStyle": {
-        "color": "green" if pnl > 0 else "red",
-        "type": "dashed",
-        "dashOffset": 5,
-      },
-    }
-    trades_lines.append([trade_line_entry, trade_line_exit])
-
-  options = {
-    "animation": False,
-    "dataset": { "source": dataset },
-    "tooltip": {
-      "trigger": 'axis',
-      "axisPointer": {
-        "type": 'line'
-      }
-    },
-    "toolbox": {
-      "feature": {
-        "dataZoom": {
-          "yAxisIndex": False
-        }
-      }
-    },
-    "grid": [
-      {
-        "left"  : '10%',
-        "right" : '10%',
-        "bottom": 200
-      },
-      {
-        "left"  : '10%',
-        "right" : '10%',
-        "height": 80,
-        "bottom": 80
-      }
-    ],
-    "xAxis": [
-      {
-        "type"       : 'category',
-        "boundaryGap": False,
-        "axisLine"   : { "onZero": False },
-        "splitLine"  : { "show": False },
-        "min"        : 'dataMin',
-        "max"        : 'dataMax'
-      },
-      {
-        "type"       : 'category',
-        "gridIndex"  : 1,
-        "boundaryGap": False,
-        "axisLine"   : { "onZero": False },
-        "axisTick"   : { "show": False },
-        "splitLine"  : { "show": False },
-        "axisLabel"  : { "show": False },
-        "min"        : 'dataMin',
-        "max"        : 'dataMax'
-      }
-    ],
-    "yAxis": [
-      {
-        "scale"    : True,
-        "splitArea": {
-          "show": True
-        }
-      },
-      {
-        "scale"      : True,
-        "gridIndex"  : 1,
-        "splitNumber": 2,
-        "axisLabel"  : { "show": False },
-        "axisLine"   : { "show": False },
-        "axisTick"   : { "show": False },
-        "splitLine"  : { "show": False }
-      }
-    ],
-    "dataZoom": [
-      {
-        "type"      : 'inside',
-        "xAxisIndex": [0, 1],
-        "start"     : 90,
-        "end"       : 100
-      },
-      {
-        "show"      : True,
-        "xAxisIndex": [0, 1],
-        "type"      : 'slider',
-        "bottom"    : 10,
-        "start"     : 90,
-        "end"       : 100
-      }
-    ],
-    "visualMap": {
-      "show"       : False,
-      "seriesIndex": 1,
-      "dimension"  : 6,
-      "pieces"     : [
-        {
-          "value": 1,
-          "color": "green"
-        },
-        {
-          "value": -1,
-          "color": "red"
-        }
-      ]
-    },
-    "series": [
-      {
-        "type"     : 'candlestick',
-        "itemStyle": {
-          "color"       : "green",
-          "color0"      : "red",
-          "borderColor" : "green",
-          "borderColor0": "red"
-        },
-        "encode": {
-          "x": 0,
-          "y": [1, 4, 3, 2]
-        },
-        "markPoint": {
-          "data"   : trades_markers,
-          "tooltip": {},
-        },
-        "markLine": {
-          "symbol": ["none", "none"],
-          "data"  : trades_lines,
-          "label" : { "position": "middle" }
-        },
-      },
-      {
-        "name"      : 'Volume',
-        "type"      : 'bar',
-        "xAxisIndex": 1,
-        "yAxisIndex": 1,
-        "itemStyle" : {
-          "color": 'lightblue'
-        },
-        "large" : True,
-        "encode": {
-          "x": 0,
-          "y": 5
-        }
-      },
-    ]
-  }
-
-  return options
-
-# histogram distribution
-def _dataframe_to_histogramchart_options(data: pd.Series, bins: int = 10, title: str = None):
-  if data.empty:
-    return None
-  
-  min = data.min()
-  max = data.max()
-  length = max - min + 1
-  bin_range = length / bins
-  values = [0] * bins
-  labels = []
-
-  label_min = min
-  label_max = min + bin_range - 1
-  for i in range(bins):
-    if i > 0:
-      label_min = label_max + bin_range
-      label_max = label_min + bin_range - 1
-    label = f"{round(label_min, 2)}\n{round(label_max, 2)}"
-    labels.append(label)
-
-  for val in data.values:
-    idx = int(math.floor(val - min) / bin_range)
-    values[idx] = values[idx] + 1
-
-  series = [{
-    "type": "bar",
-    "data": values,
-    "barGap": 0,
-  }]
-
-  options = {
-    "tooltip": {},
-    "xAxis"  : {
-      "data": labels,
-    },
-    "yAxis" : {},
-    "legend": {},
-    "series": series,
-  }
-
-  if title:
-    options["title"] = { "text": title }
-  
-  return options
-
-def _parse_data_for_json(x):
-  if isinstance(x, pd.Timestamp) or isinstance(x, pd.Timedelta):
-    return str(x)
-  return x
-
-def _statistics_to_json(stats: pd.Series, pnl: pd.Series):
-  result = {}
-  for i in range(len(stats.index)):
-    key = stats.index[i]
-    if not key.startswith("_"):
-      result[key] = _parse_data_for_json(stats.values[i])
-  
-  if pnl.empty:
-    return result
-  
-  profits = pnl[pnl > 0]
-  losses = pnl[pnl < 0]
-  profit_min = profits.min()
-  profit_max = profits.max()
-  profit_avg = profits.sum() / len(profits)
-  loss_min = losses.max() # I use max function because losses are negative values, same reasoning for losses_max
-  loss_max = losses.min()
-  loss_avg = losses.sum() / len(losses)
-  profits_less_then_10_count = len(profits[profits < 10])
-
-  result["Min. Profit"] = profit_min
-  result["Max. Profit"] = profit_max
-  result["Avg. Profit"] = profit_avg
-
-  result["Min. Loss"] = loss_min
-  result["Max. Loss"] = loss_max
-  result["Avg. Loss"] = loss_avg
-
-  result["Profit less then 10 USD count"] = profits_less_then_10_count
-
-  return result
-
-def _trades_to_json(data: pd.DataFrame):
-  if data.empty:
-    return None
-  
-  result = {}
-
-  for key in data.columns:
-    df = data[key].map(_parse_data_for_json)
-    result[key] = df.values.tolist()
-
-  return result
-
-def _dataframe_to_heatmapchart_options(df: pd.DataFrame):
-  first_name, second_name = df.index.names
-  first_levels = df.index.levels[0].to_list()
-  second_levels = df.index.levels[1].to_list()
-
-  df_1 = df.unstack()
-  df_1.columns = df_1.columns.droplevel()
-
-  values = df_1.values.T.tolist()
-  # values = df_1.values.tolist()
-
-  data = []
-  r = 0
-  max_value = 0
-  for row in values:
-    c = 0
-    for value in row:
-      formatted_value = '-' if math.isnan(value) else round(value, 2)
-      data.append([ r, c, formatted_value ])
-      if value > max_value:
-        max_value = value
-      c = c + 1
-    r = r + 1
-  
-  series = [
-    {
-      "name": f"{first_name} x {second_name} heatmap",
-      "type": "heatmap",
-      "data": data,
-      "label": { "show": True }
-    }
-  ]
-
-  title = f"{first_name} x {second_name}"
-
-  options = {
-    "title": { "text": title },
-    "animation": False,
-    "tooltip": {},
-    # "grid": {
-    #   "height": '50%',
-    #   "top": '10%'
-    # },
-    "xAxis": {
-      "type": "category",
-      "name": second_name,
-      "nameLocation": "center",
-      "nameGap": 30,
-      "data": second_levels,
-      "splitArea": { "show": True }
-    },
-    "yAxis": {
-      "type": "category",
-      "name": first_name,
-      "nameLocation": "center",
-      "nameGap": 30,
-      "data": first_levels,
-      "splitArea": { "show": True }
-    },
-    "visualMap": {
-      "show": False,
-      "min": 0,
-      "max": math.ceil(max_value),
-      "calculable": True,
-      "orient": 'horizontal',
-      "left": 'center',
-      "bottom": '15%',
-      "inRange": {
-        "color": _viridis_cmap_hex_colors
-      }
-    },
-    "series": series
-  }
-  return options
-
-def _parse_heatmap_series(heatmap: pd.Series):
-  if not isinstance(heatmap, pd.Series) or heatmap.empty:
-    return None
-  
-  params_combinations = combinations(heatmap.index.names, 2)
-  dataframes = [heatmap.groupby(list(dimensions)).agg("mean").to_frame(name="_Value") for dimensions in params_combinations]
-
-  options = [_dataframe_to_heatmapchart_options(df) for df in dataframes]
-  return options
+from ...pipeline.context import Context
+from ._dataframe_to_candlestickchart_options import _dataframe_to_candlestickchart_options
+from ._dataframe_to_grouped_barchart_options import _dataframe_to_grouped_barchart_options
+from ._dataframe_to_histogramchart_options import _dataframe_to_histogramchart_options
+from ._dataframe_to_scatterchart_cartesian_options import _dataframe_to_scatterchart_cartesian_options
+from ._dataseries_to_linechart_options import _dataseries_to_linechart_options
+from ._dataseries_to_stacked_barchart_options_splitted_by_zeroline import _dataseries_to_stacked_barchart_options_splitted_by_zeroline
+from ._heatmap import _parse_heatmap_series
+from ._statistics_to_json import _statistics_to_json
+from ._trades_to_json import _trades_to_json
 
 _METHODS_SOURCE_CODE = """
   function plot_echart(container, option) {
@@ -782,14 +273,12 @@ _STYLE_CODE = """
   }
 """
 
-def report_html(context: Context, parent_folder: str, file_suffix: str = "", strategy_name: str = None):
+def report_to_html(context: Context, file_suffix: str = ""):
   """
   Save the result statistics and metrics as html page report.
 
   `context` Context of the pipeline, see `strategy_tester.pipeline.context.Context`\n
-  `parent_folder` folder where to save the file\n
   `file_suffix` file name suffix to customize it's name\n
-  `strategy_name` (optional) Name of the backtested strategy
   """
   stats = context.stats
 
@@ -815,9 +304,9 @@ def report_html(context: Context, parent_folder: str, file_suffix: str = "", str
   entries_by_dow = metrics["entries_by_dow"] if "entries_by_dow" in metrics else pd.DataFrame()
   entries_by_month = metrics["entries_by_month"] if "entries_by_month" in metrics else pd.DataFrame()
 
-  # profits_losses_mean_by_hour = metrics["profits_losses_mean_by_hour"]
-  # profits_losses_mean_by_dow = metrics["profits_losses_mean_by_dow"]
-  # profits_losses_mean_by_month = metrics["profits_losses_mean_by_month"]
+  # profits_losses_mean_by_hour = metrics["profits_losses_mean_by_hour"] if "profits_losses_mean_by_hour" in metrics else pd.DataFrame()
+  # profits_losses_mean_by_dow = metrics["profits_losses_mean_by_dow"] if "profits_losses_mean_by_dow" in metrics else pd.DataFrame()
+  # profits_losses_mean_by_month = metrics["profits_losses_mean_by_month"] if "profits_losses_mean_by_month" in metrics else pd.DataFrame()
 
   profits_by_time_opened = metrics["profits_by_time_opened"] if "profits_by_time_opened" in metrics else pd.DataFrame()
   losses_by_time_opened = metrics["losses_by_time_opened"] if "losses_by_time_opened" in metrics else pd.DataFrame()
@@ -849,6 +338,7 @@ def report_html(context: Context, parent_folder: str, file_suffix: str = "", str
 
   heatmaps_chart_options = _parse_heatmap_series(heatmap)
 
+  strategy_name = context.strategy_name or context.strategy.__name__
   data_source = {
     "strategy_name": strategy_name,
     "statistics": statistics_json,
@@ -951,5 +441,5 @@ def report_html(context: Context, parent_folder: str, file_suffix: str = "", str
   """
 
   file_suffix = '_' + file_suffix if file_suffix != "" else ""
-  with open(f"{parent_folder}/report{file_suffix}.html", "w", encoding="utf-8") as file:
+  with open(f"{context.result_folder}/report{file_suffix}.html", "w", encoding="utf-8") as file:
     file.write(html)
