+++ src/strategy_tester/utils/flatten_list.py
@@ -1,0 +1,5 @@
+def flatten_list(matrix):
+  flat_list = []
+  for row in matrix:
+    flat_list += row
+  return flat_list
